performance/requirements.txt
annotated-types==0.5.0
async-timeout==4.0.3
bandit==1.7.5
boto3==1.28.49
botocore==1.31.49
cffi==1.15.1
click==8.1.7
contourpy==1.1.1
cryptography==41.0.3
cycler==0.11.0
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.37
greenlet==2.0.2
hiredis==2.2.3
importlib-resources==6.1.0
jmespath==1.0.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==9.1.0
numpy==1.26.0
packaging==23.1
pandas==2.1.0
pbr==5.11.1
Pillow==10.0.1
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.3.0
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-ulid==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==4.6.0
redis-om==0.2.1
rich==13.5.2
s3transfer==0.6.2
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
stevedore==5.1.0
tomli==2.0.1
tqdm==4.66.1
types-pyOpenSSL==23.2.0.2
types-redis==4.6.0.6
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
zipp==3.17.0
